Made adding and deleting rules more fine grained
Add will go all the way through and only no-op on rules that already exist
Deleting will only unlink the chain but keep the rules around and unreferenced
for faster firing and deleting.

+++ ipwaiter/orders/waiter.py
@@ -59,19 +59,15 @@ class Waiter:
             self._place_order(name, table, chain, parent, path, raw, opts, report)
 
     def _place_order(self, name, table, chain, parent, path, raw, opts, report):
-        # Stop if the chain exists
-        if self._iptables.exists(table, chain):
-            if report:
-                Logger.log(f"ipwaiter has already placed order: {name}")
-            return
 
         if report:
             Logger.log(f"ipwaiter is placing order: {name}")
 
         # Create the chain first
-        if not self._iptables.create(table, chain):
-            if report:
-                Logger.fatal(f"Failed to create chain: {chain} for table: {table}")
+        if not self._iptables.exists(table, chain):
+            if not self._iptables.create(table, chain):
+                if report:
+                    Logger.fatal(f"Failed to create chain: {chain} for table: {table}")
 
         # Add all of the rules for the
         reader = OrderReader(path, opts)
@@ -84,14 +80,21 @@ class Waiter:
 
             if ((raw and read_table == "raw") or
                     (not raw and read_table == "filter")):
-                if not self._iptables.add(read_table, chain, read_line):
-                    if report:
-                        Logger.fatal(f"Failed add. table {read_line}, chain {chain}, rule {read_line}")
+                # Add rule if needed
+                if not self._iptables.check_add(read_table, chain, read_line):
+                    if not self._iptables.add(read_table, chain, read_line):
+                        if report:
+                            Logger.fatal(f"Failed add. table {read_line}, chain {chain}, rule {read_line}")
 
         # Link the new chain to the parent chain
-        if not self._iptables.link(table, parent, chain):
+        if self._iptables.check_link(table, parent, chain):
             if report:
-                Logger.fatal(f"Failed to link chain: {chain} table: {table} to: {parent}")
+                Logger.log(f"ipwaiter has already placed order: {name}")
+            return
+        else:
+            if not self._iptables.link(table, parent, chain):
+                if report:
+                    Logger.fatal(f"Failed to link chain: {chain} table: {table} to: {parent}")
 
         if report:
             Logger.log(f"ipwaiter has placed order: {name}")
@@ -121,26 +124,17 @@ class Waiter:
         if report:
             Logger.log(f"ipwaiter is removing order: {name}")
 
-        # Unlink the chain first so we know its valid
-        if not self._iptables.unlink(table, parent, chain):
+        # Unlink the chain if possible
+        if self._iptables.check_link(table, parent, chain):
+            if not self._iptables.unlink(table, parent, chain):
+                if report:
+                    Logger.fatal(f"Failed to unlink chain: {chain} table: {table} from: {parent}")
+                else:
+                    return
+        else:
             if report:
-                Logger.fatal(f"Failed to unlink chain: {chain} table: {table} from: {parent}")
-            else:
-                return
-
-        # Flush the chain second
-        if not self._iptables.flush(table, chain):
-            if report:
-                Logger.fatal(f"Failed to flush chain: {chain} for table: {table}")
-            else:
-                return
-
-        # Then delete the chain
-        if not self._iptables.delete(table, chain):
-            if report:
-                Logger.fatal(f"Failed to delete chain: {chain} for table: {table}")
-            else:
-                return
+                Logger.log(f"ipwaiter has never placed order: {name}")
+            return
 
         if report:
             Logger.log(f"ipwaiter has removed order: {name}")
@@ -167,38 +161,41 @@ class Waiter:
 
         Logger.log("Hired ipwaiter")
 
-    def fire_waiter(self):
+    def fire_waiter(self, destroy):
         Logger.log("Firing old ipwaiter")
 
         orders = []
-        for order in os.listdir(self._order_dir):
-            abspath = utils.to_absolute_path(self._order_dir, order)
-            if os.path.isfile(abspath) and abspath.endswith(".order"):
-                base_order = os.path.basename(abspath)
-                base_order = re.sub("\.order$", "", base_order)
-                orders.append(base_order)
+        if destroy:
+            for order in os.listdir(self._order_dir):
+                abspath = utils.to_absolute_path(self._order_dir, order)
+                if os.path.isfile(abspath) and abspath.endswith(".order"):
+                    base_order = os.path.basename(abspath)
+                    base_order = re.sub("\.order$", "", base_order)
+                    orders.append(base_order)
 
         # Delete all not raw
-        self._delete_order(("input", *orders), raw=False, report=False)
-        self._delete_order(("forward", *orders), raw=False, report=False)
-        self._delete_order(("output", *orders), raw=False, report=False)
+        if destroy:
+            self._delete_order(("input", *orders), raw=False, report=False)
+            self._delete_order(("forward", *orders), raw=False, report=False)
+            self._delete_order(("output", *orders), raw=False, report=False)
 
-        # Delete all raw
-        self._delete_order(("output", *orders), raw=True, report=False)
+            # Delete all raw
+            self._delete_order(("output", *orders), raw=True, report=False)
 
         # Delete the order chains
         self._iptables.flush("filter", "input_orders")
         self._iptables.flush("filter", "forward_orders")
         self._iptables.flush("filter", "output_orders")
-        self._iptables.delete("filter", "input_orders")
-        self._iptables.delete("filter", "forward_orders")
-        self._iptables.delete("filter", "output_orders")
-
         self._iptables.flush("raw", "output_orders")
-        self._iptables.delete("raw", "output_orders")
+
+        if destroy:
+            self._iptables.delete("filter", "input_orders")
+            self._iptables.delete("filter", "forward_orders")
+            self._iptables.delete("filter", "output_orders")
+            self._iptables.delete("raw", "output_orders")
 
         Logger.log("Fired ipwaiter")
 
     def rehire_waiter(self, opts):
-        self.fire_waiter()
+        self.fire_waiter(destroy=False)
         self.hire_waiter(opts)
